+++ packages/fiori-app-sub-generator/test/unit/fiori-app-generator/fioriAppGenerator-lifecycle1.test.ts
@@ -834,7 +834,9 @@ describe('Test FioriAppGenerator', () => {
                 backendSystem: {
                     name: 'testSystemName',
                     url: 'http:/s4hsystem:1234',
-                    authenticationType: 'reentranceTicket'
+                    authenticationType: 'reentranceTicket',
+                    systemType: 'AbapCloud',
+                    connectionType: 'abap_catalog'
                 }
             }
         };
@@ -868,7 +870,9 @@ describe('Test FioriAppGenerator', () => {
                         backendSystem: {
                             authenticationType: 'reentranceTicket',
                             name: 'testSystemName',
-                            url: 'http:/s4hsystem:1234'
+                            url: 'http:/s4hsystem:1234',
+                            systemType: 'AbapCloud',
+                            connectionType: 'abap_catalog'
                         },
                         serviceProvider: {} as ServiceProvider
                     }