+++ packages/fiori-app-sub-generator/test/unit/fiori-app-generator/fioriAppGenerator-lifecycle2.test.ts
@@ -409,7 +409,9 @@ describe('Test FioriAppGenerator', () => {
                         url: 'http://mockhost:1234',
                         client: '010',
                         username: 'mockUser',
-                        password: 'mockPassword'
+                        password: 'mockPassword',
+                        systemType: 'OnPrem',
+                        connectionType: 'abap_catalog'
                     },
                     serviceProvider: {} as ServiceProvider
                 },