+++ packages/ui-service-sub-generator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ux/ui-service-sub-generator
 
+## 0.1.130
+
+### Patch Changes
+
+- Updated dependencies [ba58398]
+    - @sap-ux/system-access@0.6.33
+    - @sap-ux/fiori-generator-shared@0.13.44
+    - @sap-ux/telemetry@0.6.48
+    - @sap-ux/ui-service-inquirer@0.2.67
+    - @sap-ux/inquirer-common@0.9.17
+
 ## 0.1.129
 
 ### Patch Changes